requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
temporalio>=1.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# AWS Bedrock
boto3>=1.34.0
langchain>=0.1.0
langchain-aws>=0.1.0
langchain-community>=0.0.20

# MCP Server
mcp>=0.9.0
httpx>=0.26.0

# Utilities
requests>=2.31.0
